% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addformulabars.R
\name{AddFormulaBars}
\alias{AddFormulaBars}
\title{Add random effects terms to a model formula}
\usage{
AddFormulaBars(frml, data)
}
\arguments{
\item{frml}{A model \code{\link{formula}}.}

\item{data}{A \code{data.frame} containing the variables in
\code{frml}.}
}
\description{
Replaces factor variables in a \code{\link{formula}} with their
representation as a random effect (e.g. a factor \code{f} becomes
\code{(1|f)}) as done for formulas in e.g. the \code{lme4} and
\code{nlme} packages.
}
\details{
A formula with one factor \code{f} and one continuous
    variable \code{c} in the formula
}
\examples{
dat <- data.frame(c1 = 1:3, c2 = rnorm(3), f1 = factor(1:3), f2 = ordered(1:3))
frml <- ~c1+c2+f1+f2
AddFormulaBars(frml, dat)
}
