% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{AllVariablesNames}
\alias{AllVariablesNames}
\title{\code{AllVariablesNames}}
\usage{
AllVariablesNames(formula)
}
\arguments{
\item{formula}{A \code{\link{formula}}.}
}
\description{
Find the names of the variables (including those in dataframes) in a formula.
}

