% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocoerceclass.R
\name{AutoCoerceClass}
\alias{AutoCoerceClass}
\title{Automatically coerces a vector into an appropriate type.}
\usage{
AutoCoerceClass(x, stringsAsFactors = TRUE, max.value.labels = 12)
}
\arguments{
\item{x}{A \code{vector}.}

\item{stringsAsFactors}{logical: should character vectors be
converted to factors?}

\item{max.value.labels}{logical: only variables with value labels
and at most this many unique values will be converted to factors.
Ignored if \code{stringsAsFactors} is \code{FALSE}.}
}
\description{
Converts it to \code{\link{character}}, and then uses pattern
matching to work out the most appropriate type.
}
