% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.r
\name{Calibrate.print}
\alias{Calibrate.print}
\title{\code{Calibrate.print}}
\usage{
Calibrate.print(weight, ...)
}
\arguments{
\item{weights}{A vector of weights.}
}
\value{
character string
}
\description{
Print method for calibrate weights
}
