% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{CalibrateWeight}
\alias{CalibrateWeight}
\title{\code{CalibrateWeight}}
\usage{
CalibrateWeight(weights, strata = NULL)
}
\arguments{
\item{weights}{A vector of weights.}

\item{strata}{Strata, to perform the weight-calibration within.}
}
\value{
vector
}
\description{
Calibrates a weight to sum to the Effective Sample Size.
}
