% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{CheckCorrelationMatrix}
\alias{CheckCorrelationMatrix}
\title{CheckCorrelationMatrix}
\usage{
CheckCorrelationMatrix(correlations, variable.names = colnames(correlations))
}
\arguments{
\item{correlations}{A correlation matrix.}

\item{variable.names}{Variable names.}
}
\description{
Throws an error if there is linear dependence, or NAs, or 1s in the correlation matrix.
}
