% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{CheckForPositiveVariance}
\alias{CheckForPositiveVariance}
\title{CheckForPositiveVariance}
\usage{
CheckForPositiveVariance(x, variable.names = names(x))
}
\arguments{
\item{x}{A \code{\link{matrix}} or \code{\link{data.frame}}.}

\item{variable.names}{Variable names.}
}
\description{
Throws an error if any variables cotnain a standard deviation of 0.
}
