% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{CheckForUniqueVariableNames}
\alias{CheckForUniqueVariableNames}
\title{CheckForUniqueVariableNames}
\usage{
CheckForUniqueVariableNames(formula)
}
\arguments{
\item{formula}{A \code{\link{formula}}.}
}
\description{
Checks that the same variable name is not used twice in a formula.
}

