% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{CleanBackticks}
\alias{CleanBackticks}
\title{\code{CleanBackticks}}
\usage{
CleanBackticks(nms)
}
\arguments{
\item{nms}{Vector of variable names.}
}
\description{
Removes extra backticks and unescapes original backticks
in variable names.
}

