% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingdata.R
\name{CleanSubset}
\alias{CleanSubset}
\title{\code{CleanSubset}
Takes a QSubset variable and turns it into a logical vector with no missing values}
\usage{
CleanSubset(subset, n.total)
}
\arguments{
\item{subset}{A QSubset variable from Displayr or Q.}

\item{n.total}{The total number of observations.}
}
\description{
\code{CleanSubset}
Takes a QSubset variable and turns it into a logical vector with no missing values
}
