% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{DataFormula}
\alias{DataFormula}
\title{\code{DataFormula}}
\usage{
DataFormula(formula)
}
\arguments{
\item{formula}{An object of class \code{\link{formula}}.}
}
\description{
The formula to be used with extracted data. Same as the original formula except when
a dataframe variable is referred to, in which case the dataframe variable is converted to a normal
variable by adding backticks around it.
}

