% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{EffectiveSampleSize}
\alias{EffectiveSampleSize}
\title{\code{EffectiveSampleSize}}
\usage{
EffectiveSampleSize(weights)
}
\arguments{
\item{weights}{A vector of weights.}
}
\value{
numeric
}
\description{
Uses Kish's formula for computing an effective sample size.
}
\references{
Kish, Leslie (1965). Survey Sampling. New York: Wiley.
}
