% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingdata.R
\name{ErrorIfInfinity}
\alias{ErrorIfInfinity}
\title{Returns an error if any column(s) contain +/- ininity.}
\usage{
ErrorIfInfinity(data)
}
\arguments{
\item{data}{A \code{data.frame}.}
}
\description{
Returns an error if any column(s) contain +/- ininity.
}
