% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimationdata.R
\name{EstimationData}
\alias{EstimationData}
\title{\code{EstimationData} Selects the data from a data frame for estimation.
Conducts imputation if necessary.}
\usage{
EstimationData(formula, data, subset = NULL, weights = NULL,
  missing = "Exclude cases with missing data", m = 1, seed = 12321)
}
\arguments{
\item{formula}{An object of class \code{\link{formula}} (or one that can be
coerced to that class): a symbolic description of the model to be fitted.
The details of type specification are given under \sQuote{Details}.}

\item{data}{A \code{\link{data.frame}}.}

\item{subset}{An optional vector specifying a subset of observations to be
used in the fitting process, or, the name of a variable in \code{data}. It
may not be an expression. \code{subset} may not}

\item{weights}{An optional vector of sampling weights, or, the name or, the
name of a variable in \code{data}. It may not be an expression.}

\item{missing}{How missing data is to be treated in the regression. Options
are: \code{"Error if missing data"}, \code{"Exclude cases with missing
data"}, \code{"Use partial data"}, \code{"Use partial data (pairwise
correlations)"}, \code{"Imputation (replace missing values with
estimates)"}, and  \code{"Multiple imputation"}.}

\item{m}{Number of imputation samples.}

\item{seed}{The random number seed used in the imputation.}
}
\description{
\code{EstimationData} Selects the data from a data frame for estimation.
Conducts imputation if necessary.
}

