% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimationdata.R
\name{EstimationDataFormula}
\alias{EstimationDataFormula}
\title{\code{EstimationDataFormula}}
\usage{
EstimationDataFormula(formula)
}
\arguments{
\item{formula}{An object of class \code{\link{formula}}.}
}
\description{
The formula to be used with estimation data. Same as the original formula except when
a dataframe variable is referred to, in which case the dataframe variable is converted to a normal
variable by adding backticks around it.
}

