% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{GetTidyTwoDimensionalArray}
\alias{GetTidyTwoDimensionalArray}
\title{\code{GetTidyTwoDimensionalArray}}
\usage{
GetTidyTwoDimensionalArray(
  x,
  row.names.to.remove = NULL,
  column.names.to.remove = NULL
)
}
\arguments{
\item{x}{The data that is being analyzed}

\item{row.names.to.remove}{A vector of the row labels to remove.}

\item{column.names.to.remove}{A vector of the column labels to remove.}
}
\description{
Checks that an array is two dimensional and tidies if appropriate (assuming it is a Q table).
}
