% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{HasOutcome}
\alias{HasOutcome}
\title{\code{HasOutcome}}
\usage{
HasOutcome(formula)
}
\arguments{
\item{formula}{A \code{\link{formula}}.}
}
\value{
logical
}
\description{
Checking if the formula contains an outcome (varib)i.e., dependent variable).
}

