% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{HasSubset}
\alias{HasSubset}
\title{\code{HasSubset}}
\usage{
HasSubset(subset)
}
\arguments{
\item{subset}{The filter used to filter data in a model.}
}
\value{
true if the subset contains information
}
\description{
Checks that the subset contains data.
}

