% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{IsCount}
\alias{IsCount}
\title{\code{IsCount}}
\usage{
IsCount(x)
}
\arguments{
\item{x}{A variable or text string describing a family (e.g., "Poisson").}
}
\value{
logical.
}
\description{
Checks of data, or, a model description, counts or represents counts.
}

