% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ismissing.R
\name{IsMissing}
\alias{IsMissing}
\title{Identifies missing values in character objects}
\usage{
IsMissing(x)
}
\arguments{
\item{x}{A \code{vector}.}
}
\description{
Looks for common ways of expressing values, such as \code{NA},
\code{"NA"}, and \code{""}.
}
