% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergedatasetsbycase.R
\name{MergeDataSetsByCase}
\alias{MergeDataSetsByCase}
\title{Merge Data Sets by Case}
\usage{
MergeDataSetsByCase(
  data.set.names,
  merged.data.set.name = NULL,
  select.what.to.match.by = "Automatically",
  match.by.variable.names = TRUE,
  match.by.variable.labels = TRUE,
  match.by.value.labels = TRUE,
  min.match.percentage = 100,
  variables.to.combine = NULL,
  variables.to.not.combine = NULL,
  variables.to.omit = NULL,
  include.merged.data.set.in.output = FALSE,
  prioritize.early.data.sets = TRUE,
  use.first.value.label = TRUE,
  ignore.case = TRUE,
  ignore.special.characters = TRUE,
  data.sets.whose.variables.are.kept = seq_along(data.set.names),
  required.data.sets = NULL
)
}
\arguments{
\item{data.set.names}{A character vector of names of data sets from the
Displayr cloud drive to merge (if run from Displayr) or file paths of local
data sets.}

\item{merged.data.set.name}{A string of the name of the merged data set in
the Displayr cloud drive (if run from Displayr) or the local file path of
the merged data set. If NULL, the data set is not written.}

\item{select.what.to.match.by}{Either "Automatically" or "Manually". If
the former is chosen, the metadata to match by is chosen automatically,
whereas if the latter is chosen, the metadata to match by is specified
by setting the flags \code{match.by.variable.names},
\code{match.by.variable.labels} and match.by.value.labels.}

\item{match.by.variable.names}{Whether to match using variable names.}

\item{match.by.variable.labels}{Whether to match using variable names.}

\item{match.by.value.labels}{Whether to match using value labels of
categorical variables.}

\item{min.match.percentage}{To be decided, possibly a percentage.}

\item{variables.to.combine}{A character vector of pairs of comma-separated
variable names indicating which variables are to be combined together.
Ranges of variables can be specified by separating variable names by '-'.
Variables can be specified from specific data sets by appending '(x)' to
the variable name where x is the data set index.}

\item{variables.to.not.combine}{A character vector of comma-separated variable
names specifying variables that should never be combined together.
To specify variables from a specific data set, suffix variable names
with the data set index in parentheses, e.g., 'Q2(3)'.}

\item{variables.to.omit}{Character vector of variable names to omit from
the merged data set. To specify variables from a specific data set,
suffix the name with the data set index in parentheses, e.g., 'Q2(3)'.}

\item{include.merged.data.set.in.output}{Whether to include the merged data
set in the output.}

\item{prioritize.early.data.sets}{Whether earlier data sets should be
given priority instead of later data sets when determining which
names and labels to use.}

\item{use.first.value.label}{Whether to use the first label for a value if
there are multiple labels for the same value. Otherwise new values are
created for each label. If \code{prioritize.early.data.sets} is FALSE, the
last label is used instead of the first label.}

\item{ignore.case}{Ignore case when matching variable names.}

\item{ignore.special.characters}{Ignore special characters in variable names
except when digit characters appear both before and after special characters,
e.g., Q2__3, in which case all but the first special character is ignored,
i.e., Q2__3 is equivalent to Q2_3.}

\item{data.sets.whose.variables.are.kept}{An integer vector of indices of data
sets whose variables are to be kept. Any variable not in these data sets
will not be included in the merged data set.}
}
\value{
A list with the following elements:
\itemize{
  \item \code{merged.data.set} If \code{include.merged.data.set.in.output},
  is TRUE, this is a data frame of the merged data set.
  \item \code{input.data.set.metadata} A list containing metadata on the
    the input data sets such as variable names, labels etc.
  \item \code{merged.data.set.metadata} A list containing metadata on the
    the merged data set such as variable names, labels etc.
  \item \code{matched.names} A matrix where each row contains the names of
    the variables from the input data sets that have been merged together.
  \item \code{merged.names} A character vector containing the names of the
    variables in the merged data set.
  \item \code{omitted.variables} A list where each element contains the
    names of variables from an input data set that have been omitted from
    the merged data set.
  \item \code{input.value.attributes} A list containing value attributes
    from the input data sets, which are shown in the value attributes tables
    in the output.
  \item \code{is.saved.to.cloud} Whether the merged data set was saved to
    the Displayr cloud drive.
}
}
\description{
Merges multiple data sets by case where the data sets contain
 similar variables but different cases, e.g., data sets from different time
 periods.
}
