% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingdata.R
\name{MissingDataFail}
\alias{MissingDataFail}
\title{\code{MissingDataFail}}
\usage{
MissingDataFail()
}
\description{
Error thrown when missing values are present and the
    "Error if missing data" option was selected.
}
