% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingdata.R
\name{NoData}
\alias{NoData}
\title{\code{NoData}}
\usage{
NoData()
}
\description{
Error thrown when all cases contain missing data and no data is
    available for use as the "Exclude cases with missing data" option was
    selected.
}
