% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observed.R
\name{Observed}
\alias{Observed}
\alias{Observed.default}
\alias{Observed.MachineLearningEnsemble}
\title{\code{Observed}}
\usage{
Observed(x)

\method{Observed}{default}(x)

\method{Observed}{MachineLearningEnsemble}(x)
}
\arguments{
\item{x}{A model.}
}
\description{
Returns the variable used as the outcome when fitting a model.
}
\details{
Extracted for the entire data set (i.e., not just the specific values used in the fitting).
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Extracts the dependant variable from an object from its formula.

\item \code{MachineLearningEnsemble}: Extracts the dependant variable directly from an object.
}}

