% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{OutcomeName}
\alias{OutcomeName}
\title{\code{OutcomeName}}
\usage{
OutcomeName(formula)
}
\arguments{
\item{formula}{A \code{\link{formula}}.}
}
\value{
character.
}
\description{
Find the name of the outcome variable.
}

