% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{OutcomeVariable}
\alias{OutcomeVariable}
\title{\code{OutcomeVariable}}
\usage{
OutcomeVariable(formula, data)
}
\arguments{
\item{formula}{A \code{\link{formula}}.}

\item{data}{A \code{\link{data.frame}}.}
}
\value{
A vector of data.
}
\description{
Returns the outcome variable from a model.
}

