% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{PredictorNames}
\alias{PredictorNames}
\title{\code{PredictorNames}}
\usage{
PredictorNames(formula)
}
\arguments{
\item{formula}{A \code{\link{formula}}.}
}
\description{
Find the names of the predictor variables.
}

