% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.r
\name{print.Calibrate}
\alias{print.Calibrate}
\title{\code{Calibrate.print}}
\usage{
\method{print}{Calibrate}(weight, ...)
}
\arguments{
\item{weights}{A vector of weights.}
}
\value{
character string
}
\description{
Print method for calibrate weights
}
