% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{PrintDetails}
\alias{PrintDetails}
\title{\code{PrintDetails}}
\usage{
PrintDetails(x)
}
\arguments{
\item{x}{Something to be printed.}
}
\description{
A print function for error checking.
Prints its name and a \code{\link{summary}}.
}

