% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{ProcessQVariables}
\alias{ProcessQVariables}
\title{ProcessQVariables}
\usage{
ProcessQVariables(x)
}
\arguments{
\item{x}{A Q variable or a data frame containing Q variables.}
}
\description{
Performs tasks such as converting date variables to be categorical
based on their period.
}
