% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{RemoveBackticks}
\alias{RemoveBackticks}
\title{\code{RemoveBackticks}}
\usage{
RemoveBackticks(nms)
}
\arguments{
\item{nms}{Vector of variable names.}
}
\description{
Removes backticks surrounding variable names.
}
