% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingdata.R
\name{RemoveCasesWithAnyNA}
\alias{RemoveCasesWithAnyNA}
\title{\code{RemoveCasesWithAnyNA}}
\usage{
RemoveCasesWithAnyNA(x)
}
\arguments{
\item{x}{The input dataframe.}
}
\description{
Remove rows which contain NA.
}

