% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacking.R
\name{StackData}
\alias{StackData}
\title{Stack data set}
\usage{
StackData(
  input.data.set.name,
  stacked.data.set.name = NULL,
  common.labels = NULL,
  specify.by = NULL,
  variables.to.manually.stack = NULL,
  variables.to.omit = NULL
)
}
\arguments{
\item{input.data.set.name}{Name of data file to stack, either as a path to a
local file or file in the Displayr Cloud Drive}

\item{stacked.data.set.name}{Name of the stacked data file to be saved in
the Displayr Cloud Drive (if run from Displayr) or saved locally.}

\item{common.labels}{A character vector of common labels to be used to
identify variables to stack. To be identified, a set of variables to be
stacked must contain these labels, have the same prefix and suffix before
and after these labels and be contiguous.}

\item{specify.by}{To be filled}

\item{variables.to.omit}{To be filled}

\item{variables.to.stack}{To be filled}
}
\description{
Stacks variables in a data set.
}
