% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{SurroundWithBackticks}
\alias{SurroundWithBackticks}
\title{\code{SurroundWithBackticks}}
\usage{
SurroundWithBackticks(names)
}
\arguments{
\item{names}{A vector of names.}
}
\description{
Surrounds names with backticks, after removing existing backticks,
provided there is a $ in the name.
}

