% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{ValidateNewData}
\alias{ValidateNewData}
\title{ValidateNewData}
\usage{
ValidateNewData(object, newdata = NULL)
}
\arguments{
\item{object}{A \code{MachineLearning} or \code{Regression} object.}

\item{newdata}{Optionally, a data frame including the variables used to fit the model.}
}
\value{
A \code{data.frame} that has the appropriate variables to predict new outcome
        values for a provided model in \code{object}.
}
\description{
Checks the model provided in \code{object} either has valid data to make predictions
or that the data provided in \code{newdata} is valid to use for predictions.
}
