% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{WeightedSurveyDesign}
\alias{WeightedSurveyDesign}
\title{weightedSurveyDesign}
\usage{
WeightedSurveyDesign(data, weights)
}
\arguments{
\item{data}{The data frame.}

\item{weights}{Weights The sampling weights.}
}
\description{
weightedSurveyDesign
}
