% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{Calibrate}
\alias{Calibrate}
\title{\code{Calibrate}}
\usage{
Calibrate(categorical.variables = NULL, categorical.targets = NULL,
  numeric.variables = NULL, numeric.targets = NULL, lower = NA,
  upper = NA, trim.iterations = 20, subset = NULL,
  input.weight = NULL)
}
\arguments{
\item{categorical.variables}{An optional list or data frame of categorical adjustment variables.}

\item{categorical.targets}{The target probabilities for each category listed in \code{categorical.variables}}

\item{numeric.variables}{An optional list or data frame of categorical adjustment variables.}

\item{numeric.targets}{the target mean for each numeric variable in numeric.variables}

\item{lower}{A lower bound weight value (not guaranteed to be achieved).}

\item{upper}{An upper bound weight value (not guaranteed to be achieved).}

\item{trim.iterations}{The number of times to run the trim loop over the final weightings}

\item{subset}{A logical vector indicating which subset of cases should be used to create the weight}

\item{input.weight}{An optional weight variable; if supplied, the created weight is created to be as close
to this input.weight as possible}
}
\value{
numeric A vector of weights
}
\description{
Creates a sampling weight based on raking/calibration.
}
