% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.r
\name{Calibrate}
\alias{Calibrate}
\title{\code{Calibrate}}
\usage{
Calibrate(formCategorical, categorical.targets, formNumeric,
  numeric.targets, lower, upper, input.weight, trim.iterations)
}
\arguments{
\item{formCategorical}{Zero or more categorical sample vectors that are to be used to weight the data}

\item{categorical.targets}{The target probabilities for each category listed in formCategorical}

\item{formNumeric}{the numeric variables associated with each sample that are to be used to create the output weights}

\item{numeric.targets}{the target probabilities for each numeric variable in formNumeric}

\item{lower}{The lower bound for weighting of an individual sample}

\item{upper}{The upper bound for weighting of an individual sample}

\item{input.weight}{The sample size}

\item{trim.iterations}{The number of times to run the trim loop over the final weightings}
}
\value{
numeric A vector of weights
}
\description{
\code{Calibrate}
}
