% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{CheckPredictionVariables}
\alias{CheckPredictionVariables}
\title{\code{CheckPredictionVariables}}
\usage{
CheckPredictionVariables(object, newdata)
}
\arguments{
\item{object}{A \code{SupportVectorMachine} object.}

\item{newdata}{Optionally, a data frame including the variables used to fit the model.
If omitted, the actual data used to fit the model is used (before any filtering).}
}
\description{
Verifies that newdata is consistent with data used to used to fit a model.  newdata must contain a
superset of the variables used to fit the model or an error results. If a factor variable of
newdata contains fewer levels than the factor used for fitting, the levels are expanded.  If a factor
variable contains more levels than the factor used for fitting, a warning is given.  Returns newdata
with factor levels set to those those used for fitting and NA for new levels.
}

