% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{CheckPredictionVariables}
\alias{CheckPredictionVariables}
\title{\code{CheckPredictionVariables}}
\usage{
CheckPredictionVariables(object, newdata)
}
\arguments{
\item{object}{A \code{SupportVectorMachine} object.}

\item{newdata}{A data frame including the variables used to fit the model.
If omitted, the actual data used to fit the model is used (before any filtering).}
}
\description{
Verifies that newdata is consistent with data used to used to fit a model.  newdata must contain a
superset of the variables used to fit the model or an error results. If a factor variable contains
more levels than the factor used for fitting, a warning is given and instances with such new factor
levels are set to NA in the returned data frame.  The levels of the returned data frame are set to those
of the fitted model.
}

