% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinevariablesetsasbinary.R
\name{CombineVariableSetsAsBinary}
\alias{CombineVariableSetsAsBinary}
\title{CombineVariableSetsAsBinary}
\usage{
CombineVariableSetsAsBinary(variable.set.list, compute.for.incomplete = TRUE)
}
\arguments{
\item{variable.set.list}{A list containing one or more variable sets which should be
Nominal, Ordinal, Nominal/Ordinal - Multi, Binary - Multi, or Binary - Multi (Compact)}

\item{compute.for.incomplete}{A boolean value. If \code{FALSE}, cases with any missing data
will have a missing vlaue. If \code{TRUE}, only cases whose data is entirely missing will
be assigned a missing value.}
}
\description{
Combines a list of variable sets to binary variables, matching categories between them.
}
