% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{DataFormula}
\alias{DataFormula}
\title{\code{DataFormula}}
\usage{
DataFormula(formula)
}
\arguments{
\item{formula}{An object of class \code{\link{formula}}.}
}
\description{
Modifies formula so that any variables that refer to dataframes
with a dollar sign are surrounded by backticks. Any such variables that already
contain backticks will have those backticks removed.
}

