% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimationdata.R
\name{EstimationDataTemplate}
\alias{EstimationDataTemplate}
\title{Extract dataframe information for a template}
\usage{
EstimationDataTemplate(x, outcome.name)
}
\arguments{
\item{x}{A \code{data.frame} containing the data to be templated.}

\item{outcome.variable}{An optional name of the outcome variable. Should be a character(1L) string.
If specified it will be set as the \code{"outcome.name"} attribute in the
returned list (see below). If not specified, then the attribute will
be set to \code{NA_character_}.}
}
\value{
A list of lists. The list will have an attributes called \code{"outcome.name"} which will be
        set to the value of the \code{outcome.variable} argument or \code{NA_character_} if not
        provided. Each sublist contains information about each variable from the input
        \code{data.frame} with the following elements:
\itemize{
   \item type: The type of the variable (numeric or factor)
   \item label: The label of the variable
   \item name: The name of the variable
   \item questiontype: The question type of the variable
   \item question: The question of the variable
   \item dataset: The dataset of the variable
   \item levels: If variable is a factor, the levels of the variable, otherwise not present
   \item observed.levels: If variable is a factor, the observed levels of the variable, otherwise not present
   \item has.unobserved.levels: Boolean indicating whether the variable has unobserved levels
   \item ordered: If variable is a factor, whether the factor is ordered, otherwise not present
   \item default.value: The default value of the variable, the minimum if numeric or the first (observed)
                        level if factor
}
}
\description{
Inspects the input dataframe and creates a list of lists. Each list element takes the name
             of the variable and contains information about the variable with a sublist. Such information
             includes the type of the variable (numeric or factor), and other metadata attributes such as
             the label of the variable, the name of the variable, question type of the variable,
             the dataset it originates from and others.
}
