% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{GetData}
\alias{GetData}
\title{\code{GetData}}
\usage{
GetData(formula, data, auxiliary.data)
}
\arguments{
\item{formula}{A \code{\link{formula}}.}

\item{data}{A \code{\link{data.frame}}.}

\item{auxiliary.data}{A \code{\link{data.frame}} containing additional variables to be used in imputation (if required).}
}
\value{
character.
}
\description{
Extracts data from the environment of the formula, if it has not
been provided as an argument.
}
