% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergedatasetsbycase.R
\name{MergeDataSetsByCase}
\alias{MergeDataSetsByCase}
\title{Merge Data Sets by Case}
\usage{
MergeDataSetsByCase(
  data.set.names,
  merged.data.set.name = NULL,
  auto.select.what.to.match.by = TRUE,
  match.by.variable.names = TRUE,
  match.by.variable.labels = TRUE,
  match.by.value.labels = TRUE,
  ignore.case = TRUE,
  ignore.non.alphanumeric = TRUE,
  min.match.percentage = 90,
  variables.to.combine = NULL,
  variables.to.not.combine = NULL,
  variables.to.keep = NULL,
  variables.to.omit = NULL,
  include.merged.data.set.in.output = FALSE,
  when.multiple.labels.for.one.value = "Create new values for the labels",
  use.names.and.labels.from = "First data set",
  data.sets.whose.variables.are.kept = seq_along(data.set.names),
  min.value.label.match.percentage = 90
)
}
\arguments{
\item{data.set.names}{A character vector of names of data sets from the
Displayr cloud drive to merge (if run from Displayr) or file paths of local
data sets.}

\item{merged.data.set.name}{A string of the name of the merged data set in
the Displayr cloud drive (if run from Displayr) or the local file path of
the merged data set. If NULL, the data set is not written.}

\item{auto.select.what.to.match.by}{If TRUE, the metadata to match by is
chosen automatically, whereas if FALSE, the metadata to match by is
specified by setting the flags \code{match.by.variable.names},
\code{match.by.variable.labels} and match.by.value.labels.}

\item{match.by.variable.names}{Whether to match using variable names.}

\item{match.by.variable.labels}{Whether to match using variable names.}

\item{match.by.value.labels}{Whether to match using value labels of
categorical variables.}

\item{ignore.case}{Ignore case when matching text.}

\item{ignore.non.alphanumeric}{Ignore non-alphanumeric characters when
matching text except when numeric characters appear both before and after
non-alphanumeric characters e.g., "24 - 29", in which case the characters
are still ignored but the separation between the numbers is noted.}

\item{min.match.percentage}{To be decided, possibly a percentage.}

\item{variables.to.combine}{A character vector of comma-separated
variable names indicating which variables are to appear together.
Ranges of variables can be specified by separating variable names by '-'.
Variables can be specified from specific data sets by appending '(x)' to
the variable name where x is the data set index.}

\item{variables.to.not.combine}{A character vector of comma-separated variable
names specifying variables that should never be combined together.
To specify variables from a specific data set, suffix variable names
with the data set index in parentheses, e.g., 'Q2(3)'.}

\item{variables.to.omit}{Character vector of variable names to omit from
the merged data set. To specify variables from a specific data set,
suffix the name with the data set index in parentheses, e.g., 'Q2(3)'.
Ranges of variables can be specified by separating variable names by '-'.}

\item{include.merged.data.set.in.output}{Whether to include the merged data
set in the output.}

\item{when.multiple.labels.for.one.value}{Either "Use one of the labels" or
"Create new values for the labels". When the former is chosen, the label
from the earliest/latest data set will be chosen if use.names.and.labels.from
is "First data set"/"Last data set". If the latter is chosen, new values
are generated for the extra labels.}

\item{use.names.and.labels.from}{Either "First data set" or "Last data set".
This sets the preference for either the first or last data set when
choosing which names and labels to use in the merged data set.}

\item{data.sets.whose.variables.are.kept}{An integer vector of indices of data
sets whose variables are to be kept. Any variable not in these data sets
will not be included in the merged data set.}

\item{min.value.label.match.percentage}{The minimum percentage match for
value labels to be considered the same when combining value attributes
from different variables.}

\item{varaibles.to.keep}{Character vector of variable names to keep in
the merged data set. To specify variables from a specific data set,
suffix the name with the data set index in parentheses, e.g., 'Q2(3)'.
Ranges of variables can be specified by separating variable names by '-'.}
}
\value{
A list with the following elements:
\itemize{
  \item \code{merged.data.set} If \code{include.merged.data.set.in.output},
  is TRUE, this is a data frame of the merged data set.
  \item \code{input.data.set.metadata} A list containing metadata on the
    the input data sets such as variable names, labels etc.
  \item \code{merged.data.set.metadata} A list containing metadata on the
    the merged data set such as variable names, labels etc.
  \item \code{matched.names} A matrix where each row contains the names of
    the variables from the input data sets that have been merged together.
  \item \code{merged.names} A character vector containing the names of the
    variables in the merged data set.
  \item \code{omitted.variables} A list where each element contains the
    names of variables from an input data set that have been omitted from
    the merged data set.
  \item \code{input.value.attributes} A list containing value attributes
    from the input data sets, which are shown in the value attributes tables
    in the output.
  \item \code{is.saved.to.cloud} Whether the merged data set was saved to
    the Displayr cloud drive.
}
}
\description{
Merges multiple data sets by case where the data sets contain
 similar variables but different cases, e.g., data sets from different time
 periods.
}
