% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergedatasetsbycase.R
\name{MergeDataSetsByCase}
\alias{MergeDataSetsByCase}
\title{Merge Data Sets by Case}
\usage{
MergeDataSetsByCase(
  data.set.names,
  match.by = c("Automatic", "All", "Variable names", "Variable labels",
    "Variable and value labels"),
  min.match.percentage = 100,
  manual.matches = NULL,
  variables.to.omit = NULL
)
}
\arguments{
\item{data.set.names}{A character vector of names of data sets from the
Displayr cloud drive to merge or file paths of local data sets.}

\item{match.by}{One of Automatic, All, Variable names, Variable labels,
Variable and value labels}

\item{min.match.percentage}{To be decided, possibly a percentage.}

\item{manual.matches}{A character vector of comma-separated variable names
indicating which variables are to be matched together, e.g.,
"var_name_from_data_set_1, var_name_from_data_set_2, ..."}

\item{variables.to.omit}{A list of character vectors containing names of
variables to omit from the merged data set. The vectors in the list
correspond to the input data sets.}
}
\description{
Merges multiple data sets by case where the data sets contain
 similar variables but different cases, e.g., data sets from different time
 periods.
}
