% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergedatasetsbycase.R
\name{MergeDataSetsByCase}
\alias{MergeDataSetsByCase}
\title{Merge Data Sets by Case}
\usage{
MergeDataSetsByCase(
  data.set.names,
  merged.data.set.name = NULL,
  match.by = "Automatic",
  min.match.percentage = 100,
  manual.matches = NULL,
  variables.to.omit = NULL,
  include.merged.data.set.in.output = FALSE,
  write.data.set = TRUE,
  prioritize.early.data.sets = TRUE,
  category.value.with.multiple.labels = "Use first label"
)
}
\arguments{
\item{data.set.names}{A character vector of names of data sets from the
Displayr cloud drive to merge (if run from Displayr) or file paths of local
data sets.}

\item{merged.data.set.name}{A string of the name of the merged data set in
the Displayr cloud drive (if run from Displayr) or the local file path of
the merged data set. If NULL, the data set is not written.}

\item{match.by}{One of Automatic, All, Variable names, Variable labels,
Variable and value labels}

\item{min.match.percentage}{To be decided, possibly a percentage.}

\item{manual.matches}{A character vector of comma-separated variable names
indicating which variables are to be matched together. Ranges of variables
can be specified by separating variable names by '-'. Variables can be
specified from specific data sets by appending '(x)' to the variable name
where x is the data set index.}

\item{variables.to.omit}{A list of character vectors containing names of
variables to omit from the merged data set. The vectors in the list
correspond to the input data sets.}

\item{include.merged.data.set.in.output}{Whether to include the merged data
set in the output.}

\item{write.data.set}{Whether to save the merged data set file.}

\item{prioritize.early.data.sets}{Whether earlier data sets should be
 given priority instead of later data sets when determining which
 names and labels to use.
# need to describe outputs!}

\item{category.value.with.multiple.labels}{How to deal with a single category value
with multiple labels. Either "Use first label" or "Create new values" for
each label.}
}
\description{
Merges multiple data sets by case where the data sets contain
 similar variables but different cases, e.g., data sets from different time
 periods.
}
