% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergedatasetsbyvariable.R
\name{MergeDataSetsByVariable}
\alias{MergeDataSetsByVariable}
\title{Merge Data Sets by Variable}
\usage{
MergeDataSetsByVariable(
  data.set.names,
  merged.data.set.name = NULL,
  id.variables = NULL,
  include.or.exclude.variables = rep("Include all except those manually omitted",
    length(data.set.names)),
  variables.to.include = NULL,
  variables.to.omit = NULL,
  preferred.data.set = "First data set",
  only.keep.cases.matched.to.all.data.sets = FALSE,
  include.merged.data.set.in.output = FALSE
)
}
\arguments{
\item{data.set.names}{A character vector of names of data sets from the
Displayr cloud drive to merge (if run from Displayr) or file paths of local
data sets.}

\item{merged.data.set.name}{A string of the name of the merged data set in
the Displayr cloud drive (if run from Displayr) or the local file path of
the merged data set.}

\item{id.variables}{A string containing comma-separated names of the optional
ID variables. To specify ID variables from a specific data set, suffix the
name with the data set index in parentheses, e.g., 'ID_VAR(3)'.}

\item{variables.to.include}{A character vector of variables to include,
where each element contains comma-separated variable names. To specify
variables from a specific data set, suffix the name with the data set index
in parentheses, e.g., 'Q2(3)'. Ranges of variables can be specified by
separating variable names by '-'. This parameter only applies to data sets
where variables are specified to be included.}

\item{variables.to.omit}{A character vector of variables to omit,
where each element contains comma-separated variable names. To specify
variables from a specific data set, suffix the name with the data set index
in parentheses, e.g., 'Q2(3)'. Ranges of variables can be specified by
separating variable names by '-'. This parameter only applies to data sets
where variables are specified to be omitted.}

\item{preferred.data.set}{Either "First data set" or "Last data set".
This sets the preference for either the first or last data set when
choosing the order of cases when matching with ID variables and choosing
which ID variable to retain.}

\item{only.keep.cases.matched.to.all.data.sets}{Only keep cases if they
are present in all data sets, and discard the rest.}

\item{include.merged.data.set.in.output}{Whether to include the merged data
set in the output.}

\item{include.or.omit.variables}{A character vector where each element
corresponds to an input data set, and indicates whether variables from the
input data set are to be specified in the merged data set by specifying
the variables to include ("Only include those manually specified") or the
variables to omit ("Include all except those manually omitted").}
}
\value{
A list with the following elements:
\itemize{
  \item \code{merged.data.set} If \code{include.merged.data.set.in.output},
  is TRUE, this is a data frame of the merged data set.
  \item \code{input.data.sets.metadata} A list containing metadata on the
    the input data sets such as variable names, labels etc.
  \item \code{merged.data.set.metadata} A list containing metadata on the
    the merged data set such as variable names, labels etc.
  \item \code{is.saved.to.cloud} Whether the merged data set was saved to
    the Displayr cloud drive.
}
}
\description{
Merges multiple data sets by case where the data sets contain
 similar variables but different cases, e.g., data sets from different time
 periods.
}
