% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observed.R
\name{Observed}
\alias{Observed}
\alias{Observed.default}
\title{\code{observed} The variable used as the outcome when fitting a model. .}
\usage{
Observed(x)

\method{Observed}{default}(x)
}
\arguments{
\item{x}{A model.}
}
\description{
\code{observed} The variable used as the outcome when fitting a model. .
}
\details{
Extracted for the entire data set (i.e., not just the specific values used in the fitting).
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Exctracts the dependant variable from an object from its formula.
}}

