% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.r
\name{print.Calibrate}
\alias{print.Calibrate}
\title{\code{Calibrate.print}}
\usage{
\method{print}{Calibrate}(x, ...)
}
\arguments{
\item{x}{A vector of weights.}

\item{...}{Other ignored parameters}
}
\value{
character string
}
\description{
Print method for calibrate weights
}
