% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilities.R
\name{Probabilities}
\alias{Probabilities}
\title{Probabilities}
\usage{
Probabilities(object, newdata = NULL, ...)
}
\arguments{
\item{object}{A \code{CART}, \code{MachineLearning}, \code{MachineLearningEnsemble}
or \code{Regression} object.}

\item{newdata}{Optionally, a data frame including the variables used to fit the model.
If not provided, the object$model is used instead.}

\item{...}{Optional arguments to pass to \code{predict} or other functions.}
}
\value{
A matrix of predicted probabilities for the observation to belong to each
        class label.
}
\description{
Estimates the probability of group membership for the data passed into \code{newdata} or
the data used to fit the model if \code{newdata} is not specified. Intended to be used
for the classifiers in the packages \code{flipRegression}, \code{flipMultivariates} and
\code{flipTrees}.
}
