% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitformquestions.R
\name{SplitFormQuestions}
\alias{SplitFormQuestions}
\title{\code{SplitFormQuestions}}
\usage{
SplitFormQuestions(form.data, use.names = FALSE)
}
\arguments{
\item{form.data}{The list a from Q/Displayr dropbox controls containing
variables and questions/variable sets.}

\item{use.names}{Whether to name variables in the data frame using the
variable names in Q/Displayr. Otherwise the variable label is used.}
}
\description{
Converts a list input from Q/Displayr dropbox controls which
  could contain variables and questions/variable sets and converts it to a
  dataframe.
}
