% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacking.R
\name{StackData}
\alias{StackData}
\title{Stack data set}
\usage{
StackData(
  input.data.set.name,
  stacked.data.set.name = NULL,
  automatic.common.labels = TRUE,
  common.labels = NULL,
  specify.by = "Variable",
  manual.stacking = NULL,
  variables.to.omit = NULL,
  write.data.set = TRUE,
  include.stacked.data.set.in.output = FALSE
)
}
\arguments{
\item{input.data.set.name}{Name of data file to stack, either as a path to a
local file or file in the Displayr Cloud Drive}

\item{stacked.data.set.name}{Name of the stacked data file to be saved in
the Displayr Cloud Drive (if run from Displayr) or saved locally.}

\item{common.labels}{A character vector of common labels to be used to
identify variables to stack. To be identified, a set of variables to be
stacked must contain these labels, have the same prefix and suffix before
and after these labels.}

\item{specify.by}{"Variable" or "Observation".}

\item{manual.stacking}{To be filled}

\item{variables.to.omit}{String of comma-separated variable names to omit.
Variable ranges can be specified by supplying the start and end variables
separated by a dash '-'. If the start or end variables are left out, then
the range is assumed to start from the first variable or end at the last
variable respectively. Wildcards in variable names can be specified
with an asterisk '*'.}

\item{write.data.set}{Whether to write the stacked data set.}

\item{include.stacked.data.set.in.output}{Whether to include the stacked
data set in the output.}
}
\description{
Stacks variables in a data set.
}
