% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacking.R
\name{StackData}
\alias{StackData}
\title{Stack data set}
\usage{
StackData(
  input.data.set.name,
  stacked.data.set.name = NULL,
  stack.with.common.labels = "Automatically",
  reference.variables.to.stack = NULL,
  manual.common.labels = NULL,
  specify.by = "Variable",
  manual.stacking = NULL,
  variables.to.omit = NULL,
  include.stacked.data.set.in.output = FALSE,
  include.original.case.variable = TRUE,
  include.observation.variable = TRUE
)
}
\arguments{
\item{input.data.set.name}{Name of data file to stack, either as a path to a
local file (when running locally in R) or file in the Displayr Cloud Drive
(when running in Displayr).}

\item{stacked.data.set.name}{Name of the stacked data file to be saved in
the Displayr Cloud Drive (if run from Displayr) or saved locally.}

\item{stack.with.common.labels}{\code{"Automatically"},
\code{"Using a set of variables to stack as reference"},
\code{"Using manually input common labels"},
\code{"Disabled"}.}

\item{reference.variables.to.stack}{A character vector of sets of variables
to stack to be used as a reference to generate the common labels used for
stacking. This can be a combination of comma-separated names, wildcards
and ranges. Variable ranges can be specified by supplying the start and
end variables separated by a dash '-'. If the start or end variables are
left out, then the range is assumed to start from the first variable or
end at the last variable respectively. Wildcards in variable names can be
specified with an asterisk '*'.}

\item{manual.common.labels}{A list of sets of common labels to be used to
identify variables to stack. Only used when \code{stack.with.common.labels}
is \code{"Using manually input common labels"}. To be identified, a set of
variables to be stacked must contain these labels, and have the same prefix
and suffix before and after these labels.}

\item{specify.by}{"Variable" or "Observation". See \code{manual.stacking}.}

\item{manual.stacking}{If \code{specify.by} is "Variable", this is a
character vector where each string corresponds to the names of the
variables to be stacked into a new variable. If \code{specify.by} is
"Observation", this is a character vector where each string corresponds
the names of the variables in an observation in the stacked variables.
The strings in both cases can be a combination of comma-separated names,
wildcards and ranges.
See \code{reference.variables.to.stack} for more details on the format.}

\item{variables.to.omit}{Character vector of comma-separated names of
variables to omit. Each string can be a combination of comma-separated
names, wildcards and ranges. See \code{reference.variables.to.stack} for
more details on the format.}

\item{include.stacked.data.set.in.output}{Whether to include the stacked
data set in the output object.}

\item{include.original.case.variable}{Whether to include the \code{original_case}
variable in the stacked data set.}

\item{include.observation.variable}{Whether to include the \code{observation}
variable in the stacked data set.}
}
\value{
A list with the following elements:
\itemize{
  \item \code{stacked.data.set.metadata} A list containing metadata on the
    the stacked data set such as variable names, labels etc.
  \item \code{unstackable.names} A list of character vectors containing
    names of the variables that could not be stacked using common labels due
    to mismatching types or value attributes.
  \item \code{omitted.variables} A character vector of omitted variable names.
  \item \code{omitted.stacked.variables} A character vector of omitted
    stacked variable names.
  \item \code{common.labels.list} A list of character vectors containing
    the common labels used in the stacking. The source of these common
    labels depends on the parameter \code{stack.with.common.labels} and
    are either automatically generated, extracted from reference variables
    or manually supplied.
  \item \code{is.saved.to.cloud} Whether the stacked data set was saved to
    the Displayr cloud drive.
 }
}
\description{
Stacks variables in a data set.
}
