% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyrawdata.R
\name{TidyRawData}
\alias{TidyRawData}
\title{TidyRawData}
\usage{
TidyRawData(data, as.numeric = FALSE, as.binary = FALSE, subset = NULL,
  weights = NULL, missing = "Exclude cases with missing data",
  error.if.insufficient.obs = TRUE)
}
\arguments{
\item{data}{A \code{data.frame}.}

\item{as.numeric}{If TRUE, converts factors into numeric variables.}

\item{as.binary}{If \code{TRUE}, unordered factors are represented as dummy variables.
Otherwise, they are represented as sequential integers. Only applied if \code{as.numeric == TRUE}.}

\item{subset}{An optional vector specifying a subset of observations to be
used in the fitting process, or, the name of a variable in \code{data}. It
may not be an expression. \code{subset} may not}

\item{weights}{An optional vector of sampling weights, or, the name or, the
name of a variable in \code{data}. It may not be an expression.}

\item{missing}{\code{"Error if missing data"}, \code{"Exclude cases with missing
data"} (the default, which is equivalent to 'complete.cases'), and \code{"Use partial data"}, which removes no data.}

\item{error.if.insufficient.obs}{Throw an error if there are more variables than observations.}
}
\value{
A \code{data.frame} containing the filtered raw data, which has an attribute called
\code{weights}, containing the (filtered) vector of weights.
}
\description{
Gets the raw data, tidies subset and weights, and applies the subset, dealing with any missing values along the way.
}
